feat: create assessment

+++ internal/entity/assessment_version.go
@@ -3,6 +3,6 @@ package entity
 type AssessmentVersion struct {
 	Base
 	Version      int        `json:"version"`
-	AssessmentId int        `json:"assessmentId"`
+	AssessmentId int64      `json:"assessmentId"`
 	Assessment   Assessment `json:"assessment" gorm:"foreignKey:AssessmentId"`
 }
