+++ internal/entity/base.go
@@ -3,7 +3,7 @@ package entity
 import "time"
 
 type Base struct {
-	Id        int       `gorm:"primarykey;autoIncrement;index" json:"id"`
+	Id        int64     `gorm:"primarykey;autoIncrement;index" json:"id"`
 	CreatedAt time.Time `gorm:"autoCreateTime" json:"createdAt"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime:milli" json:"updatedAt"`
 }
