+++ internal/hse/models.go
@@ -1,7 +1,7 @@
 package hse
 
 type AssessmentModel struct {
-	Id          int    `json:"id"`
+	Id          int64  `json:"id"`
 	Name        string `json:"name"`
 	Description string `json:"description"`
 }
